backend/image_categorizer/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
numpy==2.1.1
pillow==10.4.0
onnxruntime==1.19.2
sqlite-vec==0.1.6
python-multipart==0.0.9
pyjwt==2.8.1
cryptography==41.0.7
aiofiles==23.2.1
